# backend/openapi.yaml
swagger: "2.0"
info:
  title: EnerLink API
  version: "1.0.0"
  description: >
    REST API for EnerLink system (backend contract).
    Backend port: 8080, Frontend port: 3000
host: localhost:8080
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /customers:
    get:
      summary: "List of customers"
      responses:
        200:
          description: "Returns a list of customers"
          schema:
            type: array
            items:
              $ref: "#/definitions/Customer"

    post:
      summary: "Create customer"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerCreateRequest"
      responses:
        201:
          description: "Customer created"
          schema:
            type: object
            properties:
              message:
                type: string
              customer:
                $ref: "#/definitions/Customer"
        400:
          description: "Company and email are required"
  /customers/{id}:
    get:
      summary: "Customer details"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Customer found"
          schema:
            $ref: "#/definitions/Customer"
        404:
          description: "Customer not found"

    put:
      summary: "Update customer"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdateRequest"
      responses:
        200:
          description: "Customer updated"
        404:
          description: "Customer not found"

    patch:
      summary: "Activate/deactivate customer"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [active]
            properties:
              active:
                type: boolean
      responses:
        200:
          description: "Customer status updated"
        400:
          description: "'active' field is required"
        404:
          description: "Customer not found"

  # PROVIDERS
...
definitions:
...
  Address:
    type: object
    properties:
      id: { type: integer }
      street_name: { type: string }
      building_nr: { type: integer }
      apartment_nr: { type: integer }
      post_code: { type: string }
      id_city: { type: integer }
      id_district: { type: integer }
      id_country: { type: integer }
      created_at: { type: string }
  CustomerAddress:
    type: object
    properties:
      id: { type: integer }
      id_customer: { type: integer }
      id_address: { type: integer }
      address:
        $ref: "#/definitions/Address"
  Customer:
    type: object
    properties:
      id: { type: integer }
      name: { type: string }
      last_name: { type: string }
      company: { type: string }
      email: { type: string }
      phone: { type: string }
      pesel: { type: integer }
      id_document: { type: string }
      nip: { type: string }
      regon: { type: integer }
      pkwiu:
        type: object
      representative: { type: string }
      private_or_company: { type: integer }
      active: { type: boolean }
      created_at: { type: string }
      address:
        $ref: "#/definitions/CustomerAddress"
  CustomerCreateRequest:
    type: object
    required: [company, email]
    properties:
      company: { type: string }
      email: { type: string }
      nip: { type: string }
      phone: { type: string }
      address:
        type: object
        properties:
          street_name: { type: string }
          building_nr: { type: integer }
          apartment_nr: { type: integer }
          post_code: { type: string }
          id_city: { type: integer }
          id_district: { type: integer }
          id_country: { type: integer }
  CustomerUpdateRequest:
    type: object
    properties:
      company: { type: string }
      email: { type: string }
      nip: { type: string }
      phone: { type: string }
      address:
        type: object

  # PROVIDERS / OFFERS
...